Update electron package metadata

+++ make.js
@@ -526,15 +526,21 @@ w.task("chore:clean-tsc-cache", [], () => {
 w.task("chore:update-versions", [], () => {
   const version = require("./package.json").version;
   for (const pkg of ["kate-tools", "kate-core", "kate-desktop"]) {
-    const json = JSON.parse(
-      FS.readFileSync(Path.join("packages", pkg, "package.json"))
-    );
-    json.version = version;
-    FS.writeFileSync(
-      Path.join("packages", pkg, "package.json"),
-      JSON.stringify(json, null, 2)
-    );
-    console.log(`--> Updated ${pkg} to ${version}`);
+    for (const file of ["package.json", "electron-package.json"]) {
+      if (!FS.existsSync(Path.join("packages", pkg, file))) {
+        continue;
+      }
+
+      const json = JSON.parse(
+        FS.readFileSync(Path.join("packages", pkg, file))
+      );
+      json.version = version;
+      FS.writeFileSync(
+        Path.join("packages", pkg, file),
+        JSON.stringify(json, null, 2)
+      );
+      console.log(`--> Updated ${pkg} to ${version}`);
+    }
   }
 });
